+++ Changelog.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.1.0+1 (UNRELEASED)
+--------------------
+* [Improvement] Rename cli command current -> last
+* [Improvement] updated append command, added --no-edit key
+* [Improvement] improve add_message feature, new --split-by key to add multiple entries at once
 
 0.1.0 (2016-07-06)
 ------------------